Deliverable 3
The purpose of the Pewlett Hackard Analysis was to prepare for the transition of the company for when a large section of employees begin to retire.  The company is attempting to identify retiring employees by their title, department, those eligible for the mentorship program, and the roles to be filled.  Using PgAdmin, we created queries that grouped employees by:

1- Retiring employees
2- Retiring employees that removed duplicate employees who held multiple positions
3- Number of employees grouped by their respective titles
4- Employees eligible for the mentorship program